requirements.txt
python-dotenv
bs4
gradio
opencv-python
gtts
pillow
torch
transformers
gdown
fpdf
tensorflow
tensorflow-cpu